requirements.txt
python-twitter
python-dotenv
nltk
pep8
